move fake structures to common folder
